--- src/main/java/com/supriya/miniproject/repository/StudentRepository.java
@@ -1,9 +1,0 @@
-package com.supriya.miniproject.repository;
-
-import com.supriya.miniproject.entity.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
-}
