+++ src/main/java/com/swa/miniproject/MiniProjectApplication.java
@@ -1,6 +1,5 @@
-package com.supriya.miniproject;
+package com.swa.miniproject;
 
-import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
