+++ src/main/java/com/swa/miniproject/controller/StudentController.java
@@ -1,0 +1,58 @@
+package com.swa.miniproject.controller;
+
+import org.springframework.batch.core.*;
+import org.springframework.batch.core.launch.JobLauncher;
+import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
+import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
+import org.springframework.batch.core.repository.JobRestartException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@RestController
+@RequestMapping("/students")
+public class StudentController {
+
+    @Autowired
+    JobLauncher jobLauncher;
+
+    @Autowired
+    Job job;
+
+    @GetMapping("/guest")
+    public String guestStatus(){
+        return "Welcome GUEST :) :)" ;
+    }
+
+    @GetMapping("/all")
+    public String allStatus(){
+        return "Welcome ADMIN/CLIENT authority :) :)" ;
+    }
+
+    @GetMapping("/client")
+    public String userStatus(){
+        return "Welcome CLIENT authority :) :)" ;
+    }
+
+    @GetMapping("/load")
+    public BatchStatus load() throws JobParametersInvalidException, JobExecutionAlreadyRunningException, JobRestartException, JobInstanceAlreadyCompleteException {
+
+        Map<String, JobParameter> map = new HashMap<>();
+        map.put("time", new JobParameter(System.currentTimeMillis()));
+        JobParameters parameters = new JobParameters(map);
+        JobExecution jobExecution = jobLauncher.run(job, parameters);
+
+        System.out.println("JobExecution: " + jobExecution.getStatus());
+
+        System.out.println("Batch is Running...");
+        while (jobExecution.isRunning()) {
+            System.out.println("...");
+        }
+
+        return jobExecution.getStatus();
+    }
+}
