+++ src/main/java/com/swa/miniproject/entity/Student.java
@@ -1,0 +1,48 @@
+package com.swa.miniproject.entity;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.*;
+import java.time.LocalDate;
+
+@Entity
+@Setter @Getter
+public class Student {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+    private String firstName;
+    private String lastName;
+    private double gpa;
+
+    @Transient
+    private Integer age;
+
+    private LocalDate dob;
+
+    public Student(){}
+    public Student(String firstName, String lastName) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+
+    public Student(String firstName, String lastName, double gpa, LocalDate dob) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.gpa = gpa;
+        this.dob = dob;
+    }
+
+    @Override
+    public String toString() {
+        return "Student{" +
+                "id=" + id +
+                ", firstName='" + firstName + '\'' +
+                ", lastName='" + lastName + '\'' +
+                ", gpa=" + gpa +
+                ", age='" + age + '\'' +
+                ", dob=" + dob +
+                '}';
+    }
+}
