+++ src/main/java/com/swa/miniproject/repository/StudentRepository.java
@@ -1,0 +1,9 @@
+package com.swa.miniproject.repository;
+
+import com.swa.miniproject.entity.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface StudentRepository extends JpaRepository<Student, Integer> {
+}
