+++ src/main/java/com/swa/miniproject/security/JwtRequest.java
@@ -1,4 +1,4 @@
-package com.supriya.miniproject.security;
+package com.swa.miniproject.security;
 
 import java.io.Serializable;
