+++ src/main/resources/application.properties
@@ -15,13 +15,26 @@ spring.main.allow-circular-references = true
 #spring.datasource.password=password
 #spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
+
+#spring.datasource.url=jdbc:mysql://localhost:3306/csv_db?createDatabaseIfNotExist=true
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.datasource.username=root
+#spring.datasource.password=root
+
+#spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:${MYSQL_PORT:3306}/csv_db?createDatabaseIfNotExist=true
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+#spring.datasource.username=${MYSQL_USER:root}
+#spring.datasource.password=${MYSQL_PASSWORD:root}
+
+
 spring.datasource.url=jdbc:mysql://localhost:3306/csv_db?createDatabaseIfNotExist=true
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.username=root
 spring.datasource.password=root
+
 spring.batch.jdbc.initialize-schema=always
 spring.jpa.show-sql=true
 spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=create-drop
 #
 spring.batch.job.enabled=false
