+++ src/test/java/com/swa/miniproject/MiniProjectApplicationTests.java
@@ -1,4 +1,4 @@
-package com.supriya.miniproject;
+package com.swa.miniproject;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
